rename CLDR to asset-registries/LocaleData
